requirements.txt
pdfplumber==0.10.0
pdfminer-six==20221105
PyMuPDF>=1.19.6
pytesseract>=0.3.9
pdf2image>=1.16.0
opencv-python>=4.5.5
pandas>=1.4.0
transformers>=4.20.0
rapidfuzz>=2.0.0
camelot-py==0.10.1
numpy
Flask
python-dotenv
Pillow
google-generativeai
